requirements.txt
fastapi
uvicorn
easyocr
gunicorn
pandas
openpyxl
numpy
pytesseract
pillow
python-multipart
gunicorn
